Added authentication to blog

+++ Blog_API/Controllers/BlogPostController.cs
@@ -7,6 +7,9 @@ using Blog_API.Context;
 using Blog_API.Models;
 using Microsoft.EntityFrameworkCore.Diagnostics;
 using System.Threading.Tasks;
+using Microsoft.AspNetCore.Authorization;
+using System.Security.Claims;
+using Microsoft.EntityFrameworkCore.Query;
 
 
 namespace Blog_API.Controllers
@@ -22,15 +25,25 @@ namespace Blog_API.Controllers
             _context = context;
         }
 
+       
         [HttpPost]
+        [Authorize]
         public async Task<ActionResult> CreateBlog([FromBody] BlogPostModel newBlog)
         {
+            var userId = GetUserId();
+            if(userId == 0)
+            {
+                return BadRequest("nvalid user token");
+            }
+
+            newBlog.UserId = userId;
+
             _context.BlogPostTable.Add(newBlog);
             await _context.SaveChangesAsync();
             return Ok(newBlog);
         }
 
-
+       
         [HttpGet]
         public async Task<IEnumerable<BlogPostModel>> GetAllBlogs()
         {
@@ -50,8 +63,17 @@ namespace Blog_API.Controllers
             return Ok(blog);
         }
 
+        [HttpGet]
+        [Authorize]
+        public async Task<IEnumerable<BlogPostModel>> GetMyBlogs()
+        {
+            var userId = GetUserId();
+            return await _context.BlogPostTable.Where(t => t.UserId == userId).ToListAsync();            
+        }
 
+        
         [HttpPut]
+        [Authorize]
         public async Task<ActionResult<BlogPostModel>> EditBlog([FromBody] BlogPostModel blog)
         {
             var blogToEdit = await _context.BlogPostTable.FirstOrDefaultAsync(t => t.Id == blog.Id);
@@ -61,31 +83,67 @@ namespace Blog_API.Controllers
                 return NotFound("The item you're trying to edit does not exist");
             }
 
-            blogToEdit.Content = blog.Content;
+            var userId = GetUserId();
+            if (userId == 0 || userId != blogToEdit.UserId)
+                return BadRequest("You can only edit your own blogs!!");
+
+            blogToEdit.Title = blog.Title;
+            blogToEdit.Content = blog.Content;            
             
             await _context.SaveChangesAsync();
             return Ok(blogToEdit);
         }
 
 
+
+        
         [HttpDelete("{id}")]
+        [Authorize]
         public async Task<ActionResult> DeleteBlog(int id)
         {
             var blogToDelete = _context.BlogPostTable.FirstOrDefault(t => t.Id == id);
 
+            //check if null input
             if (blogToDelete == null)
             {
                 return NotFound("Item does not exist");
             }
-            if(typeof(int) != id.GetType())
+
+            //check if input type is correct
+            if (typeof(int) != id.GetType())
             {
                 return BadRequest("Wrong input type");
             }
-           
+
+            //check if user owns blog
+            var userId = GetUserId();
+            if (userId == 0 || userId != blogToDelete.UserId)
+                return BadRequest("You can only delete your own blogs");            
+
+                      
             _context.BlogPostTable.Remove(blogToDelete);
             await _context.SaveChangesAsync();
             return Ok("Item successfully deleted");
         }
+
+
+        // Helper method to get the authenticated user's ID from claims
+        private int GetUserId()
+        {
+            var identity = HttpContext.User.Identity as ClaimsIdentity;
+            if (identity == null)
+                return 0;
+
+            var userIdClaim = identity.FindFirst("id");
+            if (userIdClaim == null)
+                return 0;
+
+            if (int.TryParse(userIdClaim.Value, out int userId))
+                return userId;
+
+            return 0;
+        }
+
     }
 
 }
